requirements.txt
certifi==2016.2.28
click==6.7
colorama==0.3.9
colorlog==4.2.1
cx-Oracle==6.0b2
py==1.4.34
pytest==3.2.1
PyYAML==3.12
wincertstore==0.2
